Refactor protocols to match template module
TODO for later: check all protocols for their component rewriter -> merge multiple rewriters into one base rewriter (also add getComponentRewriter() to AbstractProtocol).
