+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_10to1_11/data/BlockEntityNames1_11.java
@@ -15,13 +15,13 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viaversion.protocols.v1_10to1_11.rewriter;
+package com.viaversion.viaversion.protocols.v1_10to1_11.data;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 import com.viaversion.viaversion.util.Key;
 
-public class BlockEntityRewriter {
+public class BlockEntityNames1_11 {
     private static final BiMap<String, String> OLD_TO_NEW_NAMES = HashBiMap.create();
 
     // Source: https://www.minecraftforum.net/forums/minecraft-java-edition/redstone-discussion-and/commands-command-blocks-and/2724507-1-11-nbt-changes-and-additions#AllTiles
