+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_10to1_11/rewriter/ItemPacketRewriter1_11.java
@@ -20,10 +20,9 @@ package com.viaversion.viaversion.protocols.v1_10to1_11.rewriter;
 import com.viaversion.viaversion.api.connection.UserConnection;
 import com.viaversion.viaversion.api.minecraft.item.Item;
 import com.viaversion.viaversion.api.protocol.remapper.PacketHandlers;
-import com.viaversion.viaversion.api.type.Type;
 import com.viaversion.viaversion.api.type.Types;
 import com.viaversion.viaversion.protocols.v1_10to1_11.Protocol1_10To1_11;
-import com.viaversion.viaversion.protocols.v1_10to1_11.rewriter.EntityIdRewriter;
+import com.viaversion.viaversion.protocols.v1_10to1_11.data.EntityNames1_11;
 import com.viaversion.viaversion.protocols.v1_9_1to1_9_3.packet.ClientboundPackets1_9_3;
 import com.viaversion.viaversion.protocols.v1_9_1to1_9_3.packet.ServerboundPackets1_9_3;
 import com.viaversion.viaversion.rewriter.ItemRewriter;
@@ -52,12 +51,13 @@ public class ItemPacketRewriter1_11 extends ItemRewriter<ClientboundPackets1_9_3
 
                         int size = wrapper.passthrough(Types.UNSIGNED_BYTE);
                         for (int i = 0; i < size; i++) {
-                            EntityIdRewriter.toClientItem(wrapper.passthrough(Types.ITEM1_8)); // Input Item
-                            EntityIdRewriter.toClientItem(wrapper.passthrough(Types.ITEM1_8)); // Output Item
+                            handleItemToClient(wrapper.user(), wrapper.passthrough(Types.ITEM1_8)); // Input Item
+                            handleItemToClient(wrapper.user(), wrapper.passthrough(Types.ITEM1_8)); // Output Item
 
                             boolean secondItem = wrapper.passthrough(Types.BOOLEAN); // Has second item
-                            if (secondItem)
-                                EntityIdRewriter.toClientItem(wrapper.passthrough(Types.ITEM1_8)); // Second Item
+                            if (secondItem) {
+                                handleItemToClient(wrapper.user(), wrapper.passthrough(Types.ITEM1_8)); // Second Item
+                            }
 
                             wrapper.passthrough(Types.BOOLEAN); // Trade disabled
                             wrapper.passthrough(Types.INT); // Number of tools uses
@@ -74,13 +74,13 @@ public class ItemPacketRewriter1_11 extends ItemRewriter<ClientboundPackets1_9_3
 
     @Override
     public Item handleItemToClient(UserConnection connection, Item item) {
-        EntityIdRewriter.toClientItem(item);
+        EntityNames1_11.toClientItem(item);
         return item;
     }
 
     @Override
     public Item handleItemToServer(UserConnection connection, Item item) {
-        EntityIdRewriter.toServerItem(item);
+        EntityNames1_11.toServerItem(item);
         if (item == null) return null;
         boolean newItem = item.identifier() >= 218 && item.identifier() <= 234;
         newItem |= item.identifier() == 449 || item.identifier() == 450;
