--- common/src/main/java/com/viaversion/viaversion/protocols/v1_11_1to1_12/metadata/MetadataRewriter1_12To1_11_1.java
@@ -1,53 +1,0 @@
-/*
- * This file is part of ViaVersion - https://github.com/ViaVersion/ViaVersion
- * Copyright (C) 2016-2024 ViaVersion and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.viaversion.viaversion.protocols.v1_11_1to1_12.metadata;
-
-import com.viaversion.viaversion.api.minecraft.entities.EntityType;
-import com.viaversion.viaversion.api.minecraft.entities.EntityTypes1_12;
-import com.viaversion.viaversion.api.minecraft.item.Item;
-import com.viaversion.viaversion.protocols.v1_11_1to1_12.Protocol1_11_1To1_12;
-import com.viaversion.viaversion.protocols.v1_9_1to1_9_3.packet.ClientboundPackets1_9_3;
-import com.viaversion.viaversion.rewriter.EntityRewriter;
-
-public class MetadataRewriter1_12To1_11_1 extends EntityRewriter<ClientboundPackets1_9_3, Protocol1_11_1To1_12> {
-
-    public MetadataRewriter1_12To1_11_1(Protocol1_11_1To1_12 protocol) {
-        super(protocol);
-    }
-
-    @Override
-    protected void registerRewrites() {
-        filter().handler((event, meta) -> {
-            if (meta.getValue() instanceof Item) {
-                meta.setValue(protocol.getItemRewriter().handleItemToClient(event.user(), meta.value()));
-            }
-        });
-
-        filter().type(EntityTypes1_12.EntityType.EVOCATION_ILLAGER).index(12).toIndex(13); // Aggressive
-    }
-
-    @Override
-    public EntityType typeFromId(int type) {
-        return EntityTypes1_12.getTypeFromId(type, false);
-    }
-
-    @Override
-    public EntityType objectTypeFromId(int type) {
-        return EntityTypes1_12.getTypeFromId(type, true);
-    }
-}
