+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_13_2to1_14/rewriter/EntityPacketRewriter1_14.java
@@ -17,40 +17,46 @@
  */
 package com.viaversion.viaversion.protocols.v1_13_2to1_14.rewriter;
 
+import com.viaversion.viaversion.api.Via;
 import com.viaversion.viaversion.api.minecraft.ClientWorld;
 import com.viaversion.viaversion.api.minecraft.Position;
+import com.viaversion.viaversion.api.minecraft.VillagerData;
 import com.viaversion.viaversion.api.minecraft.entities.EntityType;
 import com.viaversion.viaversion.api.minecraft.entities.EntityTypes1_13;
 import com.viaversion.viaversion.api.minecraft.entities.EntityTypes1_14;
+import com.viaversion.viaversion.api.minecraft.item.DataItem;
+import com.viaversion.viaversion.api.minecraft.item.Item;
 import com.viaversion.viaversion.api.minecraft.metadata.Metadata;
 import com.viaversion.viaversion.api.protocol.packet.PacketWrapper;
 import com.viaversion.viaversion.api.protocol.remapper.PacketHandlers;
-import com.viaversion.viaversion.api.type.Type;
 import com.viaversion.viaversion.api.type.Types;
 import com.viaversion.viaversion.api.type.types.version.Types1_13_2;
 import com.viaversion.viaversion.api.type.types.version.Types1_14;
 import com.viaversion.viaversion.protocols.v1_12_2to1_13.packet.ClientboundPackets1_13;
 import com.viaversion.viaversion.protocols.v1_13_2to1_14.Protocol1_13_2To1_14;
-import com.viaversion.viaversion.protocols.v1_13_2to1_14.metadata.MetadataRewriter1_14To1_13_2;
 import com.viaversion.viaversion.protocols.v1_13_2to1_14.packet.ClientboundPackets1_14;
 import com.viaversion.viaversion.protocols.v1_13_2to1_14.storage.EntityTracker1_14;
+import com.viaversion.viaversion.rewriter.EntityRewriter;
 import java.util.LinkedList;
 import java.util.List;
 
-public class EntityPacketRewriter1_14 {
+public class EntityPacketRewriter1_14 extends EntityRewriter<ClientboundPackets1_13, Protocol1_13_2To1_14> {
 
-    public static void register(Protocol1_13_2To1_14 protocol) {
-        MetadataRewriter1_14To1_13_2 metadataRewriter = protocol.get(MetadataRewriter1_14To1_13_2.class);
+    public EntityPacketRewriter1_14(Protocol1_13_2To1_14 protocol) {
+        super(protocol);
+    }
 
+    @Override
+    protected void registerPackets() {
         protocol.registerClientbound(ClientboundPackets1_13.ADD_EXPERIENCE_ORB, wrapper -> {
             int entityId = wrapper.passthrough(Types.VAR_INT);
-            metadataRewriter.tracker(wrapper.user()).addEntity(entityId, EntityTypes1_14.EXPERIENCE_ORB);
+            tracker(wrapper.user()).addEntity(entityId, EntityTypes1_14.EXPERIENCE_ORB);
         });
 
         protocol.registerClientbound(ClientboundPackets1_13.ADD_GLOBAL_ENTITY, wrapper -> {
             int entityId = wrapper.passthrough(Types.VAR_INT);
             if (wrapper.passthrough(Types.BYTE) == 1) {
-                metadataRewriter.tracker(wrapper.user()).addEntity(entityId, EntityTypes1_14.LIGHTNING_BOLT);
+                tracker(wrapper.user()).addEntity(entityId, EntityTypes1_14.LIGHTNING_BOLT);
             }
         });
 
@@ -76,7 +82,7 @@ public class EntityPacketRewriter1_14 {
                     int typeId = wrapper.get(Types.VAR_INT, 1);
 
                     EntityTypes1_13.EntityType type1_13 = EntityTypes1_13.getTypeFromId(typeId, true);
-                    typeId = metadataRewriter.newEntityId(type1_13.getId());
+                    typeId = newEntityId(type1_13.getId());
                     EntityType type1_14 = EntityTypes1_14.getTypeFromId(typeId);
 
                     if (type1_14 != null) {
@@ -94,7 +100,7 @@ public class EntityPacketRewriter1_14 {
                                 default -> typeId; // default 0 = rideable minecart
                             };
                         } else if ((type1_14.is(EntityTypes1_14.ITEM) && data > 0)
-                                || type1_14.isOrHasParent(EntityTypes1_14.ABSTRACT_ARROW)) {
+                            || type1_14.isOrHasParent(EntityTypes1_14.ABSTRACT_ARROW)) {
                             if (type1_14.isOrHasParent(EntityTypes1_14.ABSTRACT_ARROW)) {
                                 wrapper.set(Types.INT, 0, data - 1);
                             }
@@ -133,7 +139,7 @@ public class EntityPacketRewriter1_14 {
                 map(Types.SHORT); // 11 - Velocity Z
                 map(Types1_13_2.METADATA_LIST, Types1_14.METADATA_LIST); // 12 - Metadata
 
-                handler(metadataRewriter.trackerAndRewriterHandler(Types1_14.METADATA_LIST));
+                handler(trackerAndRewriterHandler(Types1_14.METADATA_LIST));
             }
         });
 
@@ -145,7 +151,7 @@ public class EntityPacketRewriter1_14 {
                 map(Types.VAR_INT);
                 map(Types.BLOCK_POSITION1_8, Types.BLOCK_POSITION1_14);
                 map(Types.BYTE);
-                handler(wrapper -> metadataRewriter.tracker(wrapper.user()).addEntity(wrapper.get(Types.VAR_INT, 0), EntityTypes1_14.PAINTING));
+                handler(wrapper -> tracker(wrapper.user()).addEntity(wrapper.get(Types.VAR_INT, 0), EntityTypes1_14.PAINTING));
             }
         });
 
@@ -161,7 +167,7 @@ public class EntityPacketRewriter1_14 {
                 map(Types.BYTE); // 6 - Pitch
                 map(Types1_13_2.METADATA_LIST, Types1_14.METADATA_LIST); // 7 - Metadata
 
-                handler(metadataRewriter.trackerAndRewriterHandler(Types1_14.METADATA_LIST, EntityTypes1_14.PLAYER));
+                handler(trackerAndRewriterHandler(Types1_14.METADATA_LIST, EntityTypes1_14.PLAYER));
             }
         });
 
@@ -180,7 +186,7 @@ public class EntityPacketRewriter1_14 {
                         metadataPacket.write(Types.VAR_INT, entityId);
                         List<Metadata> metadataList = new LinkedList<>();
                         if (tracker.clientEntityId() != entityId) {
-                            metadataList.add(new Metadata(6, Types1_14.META_TYPES.poseType, MetadataRewriter1_14To1_13_2.recalculatePlayerPose(entityId, tracker)));
+                            metadataList.add(new Metadata(6, Types1_14.META_TYPES.poseType, EntityPacketRewriter1_14.recalculatePlayerPose(entityId, tracker)));
                         }
                         metadataList.add(new Metadata(12, Types1_14.META_TYPES.optionalBlockPositionType, null));
                         metadataPacket.write(Types1_14.METADATA_LIST, metadataList);
@@ -202,7 +208,7 @@ public class EntityPacketRewriter1_14 {
                     int dimensionId = wrapper.get(Types.INT, 1);
                     clientChunks.setEnvironment(dimensionId);
                 });
-                handler(metadataRewriter.playerTrackerHandler());
+                handler(playerTrackerHandler());
                 handler(wrapper -> {
                     short difficulty = wrapper.read(Types.UNSIGNED_BYTE); // 19w11a removed difficulty from join game
                     PacketWrapper difficultyPacket = wrapper.create(ClientboundPackets1_14.CHANGE_DIFFICULTY);
@@ -239,14 +245,202 @@ public class EntityPacketRewriter1_14 {
                     List<Metadata> metadataList = new LinkedList<>();
                     metadataList.add(new Metadata(12, Types1_14.META_TYPES.optionalBlockPositionType, position));
                     if (tracker.clientEntityId() != entityId) {
-                        metadataList.add(new Metadata(6, Types1_14.META_TYPES.poseType, MetadataRewriter1_14To1_13_2.recalculatePlayerPose(entityId, tracker)));
+                        metadataList.add(new Metadata(6, Types1_14.META_TYPES.poseType, EntityPacketRewriter1_14.recalculatePlayerPose(entityId, tracker)));
                     }
                     wrapper.write(Types1_14.METADATA_LIST, metadataList);
                 });
             }
         });
 
-        metadataRewriter.registerRemoveEntities(ClientboundPackets1_13.REMOVE_ENTITIES);
-        metadataRewriter.registerMetadataRewriter(ClientboundPackets1_13.SET_ENTITY_DATA, Types1_13_2.METADATA_LIST, Types1_14.METADATA_LIST);
+        registerRemoveEntities(ClientboundPackets1_13.REMOVE_ENTITIES);
+        registerMetadataRewriter(ClientboundPackets1_13.SET_ENTITY_DATA, Types1_13_2.METADATA_LIST, Types1_14.METADATA_LIST);
+    }
+
+    @Override
+    protected void registerRewrites() {
+        filter().mapMetaType(Types1_14.META_TYPES::byId);
+        registerMetaTypeHandler(Types1_14.META_TYPES.itemType, Types1_14.META_TYPES.optionalBlockStateType, Types1_14.META_TYPES.particleType);
+
+        filter().type(EntityTypes1_14.ENTITY).addIndex(6);
+        filter().type(EntityTypes1_14.LIVING_ENTITY).addIndex(12);
+
+        filter().type(EntityTypes1_14.LIVING_ENTITY).index(8).handler((event, meta) -> {
+            float value = ((Number) meta.getValue()).floatValue();
+            if (Float.isNaN(value) && Via.getConfig().is1_14HealthNaNFix()) {
+                meta.setValue(1F);
+            }
+        });
+
+        filter().type(EntityTypes1_14.MOB).index(13).handler((event, meta) -> {
+            EntityTracker1_14 tracker = tracker(event.user());
+            int entityId = event.entityId();
+            tracker.setInsentientData(entityId, (byte) ((((Number) meta.getValue()).byteValue() & ~0x4)
+                | (tracker.getInsentientData(entityId) & 0x4))); // New attacking metadata
+            meta.setValue(tracker.getInsentientData(entityId));
+        });
+
+        filter().type(EntityTypes1_14.PLAYER).handler((event, meta) -> {
+            EntityTracker1_14 tracker = tracker(event.user());
+            int entityId = event.entityId();
+            if (entityId != tracker.clientEntityId()) {
+                if (meta.id() == 0) {
+                    byte flags = ((Number) meta.getValue()).byteValue();
+                    // Mojang overrides the client-side pose updater, see OtherPlayerEntity#updateSize
+                    tracker.setEntityFlags(entityId, flags);
+                } else if (meta.id() == 7) {
+                    tracker.setRiptide(entityId, (((Number) meta.getValue()).byteValue() & 0x4) != 0);
+                }
+                if (meta.id() == 0 || meta.id() == 7) {
+                    event.createExtraMeta(new Metadata(6, Types1_14.META_TYPES.poseType, recalculatePlayerPose(entityId, tracker)));
+                }
+            }
+        });
+
+        filter().type(EntityTypes1_14.ZOMBIE).handler((event, meta) -> {
+            if (meta.id() == 16) {
+                EntityTracker1_14 tracker = tracker(event.user());
+                int entityId = event.entityId();
+                tracker.setInsentientData(entityId, (byte) ((tracker.getInsentientData(entityId) & ~0x4)
+                    | ((boolean) meta.getValue() ? 0x4 : 0))); // New attacking
+                event.createExtraMeta(new Metadata(13, Types1_14.META_TYPES.byteType, tracker.getInsentientData(entityId)));
+                event.cancel(); // "Are hands held up"
+            } else if (meta.id() > 16) {
+                meta.setId(meta.id() - 1);
+            }
+        });
+
+        filter().type(EntityTypes1_14.ABSTRACT_MINECART).index(10).handler((event, meta) -> {
+            int data = meta.value();
+            meta.setValue(protocol.getMappingData().getNewBlockStateId(data));
+        });
+
+        filter().type(EntityTypes1_14.HORSE).index(18).handler((event, meta) -> {
+            event.cancel();
+
+            int armorType = meta.value();
+            Item armorItem = null;
+            if (armorType == 1) {  //iron armor
+                armorItem = new DataItem(protocol.getMappingData().getNewItemId(727), (byte) 1, (short) 0, null);
+            } else if (armorType == 2) {  //gold armor
+                armorItem = new DataItem(protocol.getMappingData().getNewItemId(728), (byte) 1, (short) 0, null);
+            } else if (armorType == 3) {  //diamond armor
+                armorItem = new DataItem(protocol.getMappingData().getNewItemId(729), (byte) 1, (short) 0, null);
+            }
+
+            PacketWrapper equipmentPacket = PacketWrapper.create(ClientboundPackets1_14.SET_EQUIPPED_ITEM, null, event.user());
+            equipmentPacket.write(Types.VAR_INT, event.entityId());
+            equipmentPacket.write(Types.VAR_INT, 4);
+            equipmentPacket.write(Types.ITEM1_13_2, armorItem);
+            try {
+                equipmentPacket.scheduleSend(Protocol1_13_2To1_14.class);
+            } catch (final Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+
+        filter().type(EntityTypes1_14.VILLAGER).index(15).handler((event, meta) -> {
+            meta.setTypeAndValue(Types1_14.META_TYPES.villagerDatatType, new VillagerData(2, getNewProfessionId(meta.value()), 0));
+        });
+
+        filter().type(EntityTypes1_14.ZOMBIE_VILLAGER).index(18).handler((event, meta) -> {
+            meta.setTypeAndValue(Types1_14.META_TYPES.villagerDatatType, new VillagerData(2, getNewProfessionId(meta.value()), 0));
+        });
+
+        filter().type(EntityTypes1_14.ABSTRACT_ARROW).addIndex(9); // Piercing level added
+
+        filter().type(EntityTypes1_14.FIREWORK_ROCKET).index(8).handler((event, meta) -> {
+            meta.setMetaType(Types1_14.META_TYPES.optionalVarIntType);
+            if (meta.getValue().equals(0)) {
+                meta.setValue(null); // https://bugs.mojang.com/browse/MC-111480
+            }
+        });
+
+        filter().type(EntityTypes1_14.ABSTRACT_SKELETON).index(14).handler((event, meta) -> {
+            EntityTracker1_14 tracker = tracker(event.user());
+            int entityId = event.entityId();
+            tracker.setInsentientData(entityId, (byte) ((tracker.getInsentientData(entityId) & ~0x4)
+                | ((boolean) meta.getValue() ? 0x4 : 0))); // New attacking
+            event.createExtraMeta(new Metadata(13, Types1_14.META_TYPES.byteType, tracker.getInsentientData(entityId)));
+            event.cancel();  // "Is swinging arms"
+        });
+
+        filter().type(EntityTypes1_14.ABSTRACT_ILLAGER).handler((event, meta) -> {
+            if (event.index() == 14) {
+                EntityTracker1_14 tracker = tracker(event.user());
+                int entityId = event.entityId();
+                tracker.setInsentientData(entityId, (byte) ((tracker.getInsentientData(entityId) & ~0x4)
+                    | (((Number) meta.getValue()).byteValue() != 0 ? 0x4 : 0))); // New attacking
+                event.createExtraMeta(new Metadata(13, Types1_14.META_TYPES.byteType, tracker.getInsentientData(entityId)));
+                event.cancel(); // "Has target (aggressive state)"
+            } else if (event.index() > 14) {
+                meta.setId(meta.id() - 1);
+            }
+        });
+
+        filter().type(EntityTypes1_14.OCELOT).removeIndex(17); // variant
+
+        // Ocelot is not tamable anymore
+        filter().type(EntityTypes1_14.OCELOT).removeIndex(16); // owner uuid
+        filter().type(EntityTypes1_14.OCELOT).removeIndex(15); // data
+
+        filter().type(EntityTypes1_14.ABSTRACT_RAIDER).addIndex(14); // celebrating
+    }
+
+    @Override
+    public void onMappingDataLoaded() {
+        mapTypes();
+        if (Via.getConfig().translateOcelotToCat()) {
+            // A better solution for this would be to despawn the ocelot and spawn a cat in its place, but that would
+            // require a lot of data tracking and is not worth the effort.
+            mapEntityType(EntityTypes1_13.EntityType.OCELOT, EntityTypes1_14.CAT);
+        }
+    }
+
+    @Override
+    public EntityType typeFromId(int type) {
+        return EntityTypes1_14.getTypeFromId(type);
+    }
+
+    private static boolean isSneaking(byte flags) {
+        return (flags & 0x2) != 0;
+    }
+
+    private static boolean isSwimming(byte flags) {
+        return (flags & 0x10) != 0;
+    }
+
+    private static int getNewProfessionId(int old) {
+        // profession -> career
+        return switch (old) {
+            case 0 -> 5; // farmer
+            case 1 -> 9; // librarian
+            case 2 -> 4; // priest ->cleric
+            case 3 -> 1; // blacksmith -> armorer
+            case 4 -> 2; // butcher
+            case 5 -> 11; // nitwit
+            default -> 0; // none
+        };
+    }
+
+    private static boolean isFallFlying(int entityFlags) {
+        return (entityFlags & 0x80) != 0;
+    }
+
+    public static int recalculatePlayerPose(int entityId, EntityTracker1_14 tracker) {
+        byte flags = tracker.getEntityFlags(entityId);
+        // Mojang overrides the client-side pose updater, see OtherPlayerEntity#updateSize
+        int pose = 0; // standing
+        if (isFallFlying(flags)) {
+            pose = 1;
+        } else if (tracker.isSleeping(entityId)) {
+            pose = 2;
+        } else if (isSwimming(flags)) {
+            pose = 3;
+        } else if (tracker.isRiptide(entityId)) {
+            pose = 4;
+        } else if (isSneaking(flags)) {
+            pose = 5;
+        }
+        return pose;
     }
 }
