+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_13_2to1_14/rewriter/PlayerPacketRewriter1_14.java
@@ -24,14 +24,12 @@ import com.viaversion.viaversion.api.Via;
 import com.viaversion.viaversion.api.minecraft.Position;
 import com.viaversion.viaversion.api.minecraft.item.Item;
 import com.viaversion.viaversion.api.protocol.remapper.PacketHandlers;
-import com.viaversion.viaversion.api.type.Type;
 import com.viaversion.viaversion.api.type.Types;
 import com.viaversion.viaversion.protocols.v1_12_2to1_13.packet.ClientboundPackets1_13;
 import com.viaversion.viaversion.protocols.v1_13_2to1_14.Protocol1_13_2To1_14;
 import com.viaversion.viaversion.protocols.v1_13_2to1_14.packet.ServerboundPackets1_14;
-import java.util.Collections;
 
-public class PlayerPackets1_14 {
+public class PlayerPacketRewriter1_14 {
 
     public static void register(Protocol1_13_2To1_14 protocol) {
         protocol.registerClientbound(ClientboundPackets1_13.OPEN_SIGN_EDITOR, new PacketHandlers() {
