--- common/src/main/java/com/viaversion/viaversion/protocols/v1_14_4to1_15/metadata/MetadataRewriter1_15To1_14_4.java
@@ -1,55 +1,0 @@
-/*
- * This file is part of ViaVersion - https://github.com/ViaVersion/ViaVersion
- * Copyright (C) 2016-2024 ViaVersion and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.viaversion.viaversion.protocols.v1_14_4to1_15.metadata;
-
-import com.viaversion.viaversion.api.minecraft.entities.EntityType;
-import com.viaversion.viaversion.api.minecraft.entities.EntityTypes1_15;
-import com.viaversion.viaversion.api.type.types.version.Types1_14;
-import com.viaversion.viaversion.protocols.v1_14_3to1_14_4.packet.ClientboundPackets1_14_4;
-import com.viaversion.viaversion.protocols.v1_14_4to1_15.Protocol1_14_4To1_15;
-import com.viaversion.viaversion.protocols.v1_14_4to1_15.rewriter.EntityPacketRewriter1_15;
-import com.viaversion.viaversion.rewriter.EntityRewriter;
-
-public class MetadataRewriter1_15To1_14_4 extends EntityRewriter<ClientboundPackets1_14_4, Protocol1_14_4To1_15> {
-
-    public MetadataRewriter1_15To1_14_4(Protocol1_14_4To1_15 protocol) {
-        super(protocol);
-    }
-
-    @Override
-    protected void registerRewrites() {
-        registerMetaTypeHandler(Types1_14.META_TYPES.itemType, Types1_14.META_TYPES.optionalBlockStateType, Types1_14.META_TYPES.particleType);
-        filter().type(EntityTypes1_15.ABSTRACT_MINECART).index(10).handler((metadatas, meta) -> {
-            int data = meta.value();
-            meta.setValue(protocol.getMappingData().getNewBlockStateId(data));
-        });
-
-        filter().type(EntityTypes1_15.LIVING_ENTITY).addIndex(12);
-        filter().type(EntityTypes1_15.WOLF).removeIndex(18);
-    }
-
-    @Override
-    public int newEntityId(final int id) {
-        return EntityPacketRewriter1_15.getNewEntityId(id);
-    }
-
-    @Override
-    public EntityType typeFromId(int type) {
-        return EntityTypes1_15.getTypeFromId(type);
-    }
-}
