--- common/src/main/java/com/viaversion/viaversion/protocols/v1_14to1_14_1/metadata/MetadataRewriter1_14_1To1_14.java
@@ -1,42 +1,0 @@
-/*
- * This file is part of ViaVersion - https://github.com/ViaVersion/ViaVersion
- * Copyright (C) 2016-2024 ViaVersion and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.viaversion.viaversion.protocols.v1_14to1_14_1.metadata;
-
-import com.viaversion.viaversion.api.minecraft.entities.EntityType;
-import com.viaversion.viaversion.api.minecraft.entities.EntityTypes1_14;
-import com.viaversion.viaversion.protocols.v1_13_2to1_14.packet.ClientboundPackets1_14;
-import com.viaversion.viaversion.protocols.v1_14to1_14_1.Protocol1_14To1_14_1;
-import com.viaversion.viaversion.rewriter.EntityRewriter;
-
-public class MetadataRewriter1_14_1To1_14 extends EntityRewriter<ClientboundPackets1_14, Protocol1_14To1_14_1> {
-
-    public MetadataRewriter1_14_1To1_14(Protocol1_14To1_14_1 protocol) {
-        super(protocol);
-    }
-
-    @Override
-    protected void registerRewrites() {
-        filter().type(EntityTypes1_14.VILLAGER).addIndex(15);
-        filter().type(EntityTypes1_14.WANDERING_TRADER).addIndex(15);
-    }
-
-    @Override
-    public EntityType typeFromId(int type) {
-        return EntityTypes1_14.getTypeFromId(type);
-    }
-}
