+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_15_2to1_16/rewriter/EntityPacketRewriter1_16.java
@@ -17,34 +17,28 @@
  */
 package com.viaversion.viaversion.protocols.v1_15_2to1_16.rewriter;
 
-import com.viaversion.nbt.tag.ByteTag;
-import com.viaversion.nbt.tag.CompoundTag;
-import com.viaversion.nbt.tag.FloatTag;
-import com.viaversion.nbt.tag.IntTag;
-import com.viaversion.nbt.tag.ListTag;
-import com.viaversion.nbt.tag.LongTag;
-import com.viaversion.nbt.tag.StringTag;
 import com.viaversion.viaversion.api.Via;
 import com.viaversion.viaversion.api.minecraft.WorldIdentifiers;
+import com.viaversion.viaversion.api.minecraft.entities.EntityType;
 import com.viaversion.viaversion.api.minecraft.entities.EntityTypes1_16;
+import com.viaversion.viaversion.api.minecraft.metadata.Metadata;
 import com.viaversion.viaversion.api.protocol.remapper.PacketHandler;
 import com.viaversion.viaversion.api.protocol.remapper.PacketHandlers;
-import com.viaversion.viaversion.api.type.Type;
 import com.viaversion.viaversion.api.type.Types;
 import com.viaversion.viaversion.api.type.types.version.Types1_14;
 import com.viaversion.viaversion.api.type.types.version.Types1_16;
 import com.viaversion.viaversion.protocols.v1_14_4to1_15.packet.ClientboundPackets1_15;
 import com.viaversion.viaversion.protocols.v1_15_2to1_16.Protocol1_15_2To1_16;
-import com.viaversion.viaversion.protocols.v1_15_2to1_16.data.AttributeMappings;
-import com.viaversion.viaversion.protocols.v1_15_2to1_16.metadata.MetadataRewriter1_16To1_15_2;
+import com.viaversion.viaversion.protocols.v1_15_2to1_16.data.Attributes1_16;
+import com.viaversion.viaversion.protocols.v1_15_2to1_16.data.DimensionRegistries1_16;
 import com.viaversion.viaversion.protocols.v1_15_2to1_16.packet.ClientboundPackets1_16;
 import com.viaversion.viaversion.protocols.v1_15_2to1_16.packet.ServerboundPackets1_16;
 import com.viaversion.viaversion.protocols.v1_15_2to1_16.storage.InventoryTracker1_16;
+import com.viaversion.viaversion.rewriter.EntityRewriter;
 import com.viaversion.viaversion.util.Key;
-import java.util.Arrays;
 import java.util.UUID;
 
-public class EntityPacketRewriter1_16 {
+public class EntityPacketRewriter1_16 extends EntityRewriter<ClientboundPackets1_15, Protocol1_15_2To1_16> {
 
     private static final PacketHandler DIMENSION_HANDLER = wrapper -> {
         WorldIdentifiers map = Via.getConfig().get1_16WorldNamesMap();
@@ -77,89 +71,13 @@ public class EntityPacketRewriter1_16 {
         wrapper.write(Types.STRING, dimensionName); // dimension
         wrapper.write(Types.STRING, outputName); // world
     };
-    public static final CompoundTag DIMENSIONS_TAG = new CompoundTag();
-    private static final String[] WORLD_NAMES = {"minecraft:overworld", "minecraft:the_nether", "minecraft:the_end"};
 
-    static {
-        ListTag<CompoundTag> list = new ListTag<>(CompoundTag.class);
-        list.add(createOverworldEntry());
-        list.add(createOverworldCavesEntry());
-        list.add(createNetherEntry());
-        list.add(createEndEntry());
-        DIMENSIONS_TAG.put("dimension", list);
+    public EntityPacketRewriter1_16(Protocol1_15_2To1_16 protocol) {
+        super(protocol);
     }
 
-    private static CompoundTag createOverworldEntry() {
-        CompoundTag tag = new CompoundTag();
-        tag.put("name", new StringTag("minecraft:overworld"));
-        tag.put("has_ceiling", new ByteTag((byte) 0));
-        addSharedOverwaldEntries(tag);
-        return tag;
-    }
-
-    private static CompoundTag createOverworldCavesEntry() {
-        CompoundTag tag = new CompoundTag();
-        tag.put("name", new StringTag("minecraft:overworld_caves"));
-        tag.put("has_ceiling", new ByteTag((byte) 1));
-        addSharedOverwaldEntries(tag);
-        return tag;
-    }
-
-    private static void addSharedOverwaldEntries(CompoundTag tag) {
-        tag.put("piglin_safe", new ByteTag((byte) 0));
-        tag.put("natural", new ByteTag((byte) 1));
-        tag.put("ambient_light", new FloatTag(0));
-        tag.put("infiniburn", new StringTag("minecraft:infiniburn_overworld"));
-        tag.put("respawn_anchor_works", new ByteTag((byte) 0));
-        tag.put("has_skylight", new ByteTag((byte) 1));
-        tag.put("bed_works", new ByteTag((byte) 1));
-        tag.put("has_raids", new ByteTag((byte) 1));
-        tag.put("logical_height", new IntTag(256));
-        tag.put("shrunk", new ByteTag((byte) 0));
-        tag.put("ultrawarm", new ByteTag((byte) 0));
-    }
-
-    private static CompoundTag createNetherEntry() {
-        CompoundTag tag = new CompoundTag();
-        tag.put("piglin_safe", new ByteTag((byte) 1));
-        tag.put("natural", new ByteTag((byte) 0));
-        tag.put("ambient_light", new FloatTag(0.1F));
-        tag.put("infiniburn", new StringTag("minecraft:infiniburn_nether"));
-        tag.put("respawn_anchor_works", new ByteTag((byte) 1));
-        tag.put("has_skylight", new ByteTag((byte) 0));
-        tag.put("bed_works", new ByteTag((byte) 0));
-        tag.put("fixed_time", new LongTag(18000));
-        tag.put("has_raids", new ByteTag((byte) 0));
-        tag.put("name", new StringTag("minecraft:the_nether"));
-        tag.put("logical_height", new IntTag(128));
-        tag.put("shrunk", new ByteTag((byte) 1));
-        tag.put("ultrawarm", new ByteTag((byte) 1));
-        tag.put("has_ceiling", new ByteTag((byte) 1));
-        return tag;
-    }
-
-    private static CompoundTag createEndEntry() {
-        CompoundTag tag = new CompoundTag();
-        tag.put("piglin_safe", new ByteTag((byte) 0));
-        tag.put("natural", new ByteTag((byte) 0));
-        tag.put("ambient_light", new FloatTag(0));
-        tag.put("infiniburn", new StringTag("minecraft:infiniburn_end"));
-        tag.put("respawn_anchor_works", new ByteTag((byte) 0));
-        tag.put("has_skylight", new ByteTag((byte) 0));
-        tag.put("bed_works", new ByteTag((byte) 0));
-        tag.put("fixed_time", new LongTag(6000));
-        tag.put("has_raids", new ByteTag((byte) 1));
-        tag.put("name", new StringTag("minecraft:the_end"));
-        tag.put("logical_height", new IntTag(256));
-        tag.put("shrunk", new ByteTag((byte) 0));
-        tag.put("ultrawarm", new ByteTag((byte) 0));
-        tag.put("has_ceiling", new ByteTag((byte) 0));
-        return tag;
-    }
-
-    public static void register(Protocol1_15_2To1_16 protocol) {
-        MetadataRewriter1_16To1_15_2 metadataRewriter = protocol.get(MetadataRewriter1_16To1_15_2.class);
-
+    @Override
+    protected void registerPackets() {
         // Spawn lightning -> Spawn entity
         protocol.registerClientbound(ClientboundPackets1_15.ADD_GLOBAL_ENTITY, ClientboundPackets1_16.ADD_ENTITY, wrapper -> {
             int entityId = wrapper.passthrough(Types.VAR_INT);
@@ -186,11 +104,11 @@ public class EntityPacketRewriter1_16 {
             wrapper.write(Types.SHORT, (short) 0); // velocity
         });
 
-        metadataRewriter.registerTrackerWithData(ClientboundPackets1_15.ADD_ENTITY, EntityTypes1_16.FALLING_BLOCK);
-        metadataRewriter.registerTracker(ClientboundPackets1_15.ADD_MOB);
-        metadataRewriter.registerTracker(ClientboundPackets1_15.ADD_PLAYER, EntityTypes1_16.PLAYER);
-        metadataRewriter.registerMetadataRewriter(ClientboundPackets1_15.SET_ENTITY_DATA, Types1_14.METADATA_LIST, Types1_16.METADATA_LIST);
-        metadataRewriter.registerRemoveEntities(ClientboundPackets1_15.REMOVE_ENTITIES);
+        registerTrackerWithData(ClientboundPackets1_15.ADD_ENTITY, EntityTypes1_16.FALLING_BLOCK);
+        registerTracker(ClientboundPackets1_15.ADD_MOB);
+        registerTracker(ClientboundPackets1_15.ADD_PLAYER, EntityTypes1_16.PLAYER);
+        registerMetadataRewriter(ClientboundPackets1_15.SET_ENTITY_DATA, Types1_14.METADATA_LIST, Types1_16.METADATA_LIST);
+        registerRemoveEntities(ClientboundPackets1_15.REMOVE_ENTITIES);
 
         protocol.registerClientbound(ClientboundPackets1_15.RESPAWN, new PacketHandlers() {
             @Override
@@ -216,8 +134,8 @@ public class EntityPacketRewriter1_16 {
                 map(Types.UNSIGNED_BYTE); //  Gamemode
                 handler(wrapper -> {
                     wrapper.write(Types.BYTE, (byte) -1); // Previous gamemode, set to none
-                    wrapper.write(Types.STRING_ARRAY, Arrays.copyOf(WORLD_NAMES, WORLD_NAMES.length)); // World list - only used for command completion
-                    wrapper.write(Types.NAMED_COMPOUND_TAG, DIMENSIONS_TAG.copy()); // Dimension registry
+                    wrapper.write(Types.STRING_ARRAY, DimensionRegistries1_16.getWorldNames()); // World list - only used for command completion
+                    wrapper.write(Types.NAMED_COMPOUND_TAG, DimensionRegistries1_16.getDimensionsTag()); // Dimension registry
                 });
                 handler(DIMENSION_HANDLER); // Dimension
                 map(Types.LONG); // Seed
@@ -243,7 +161,7 @@ public class EntityPacketRewriter1_16 {
             for (int i = 0; i < size; i++) {
                 // Attributes have been renamed and are now namespaced identifiers
                 String key = wrapper.read(Types.STRING);
-                String attributeIdentifier = AttributeMappings.attributeIdentifierMappings().get(key);
+                String attributeIdentifier = Attributes1_16.attributeIdentifierMappings().get(key);
                 if (attributeIdentifier == null) {
                     attributeIdentifier = Key.namespaced(key);
                     if (!Key.isValid(attributeIdentifier)) {
@@ -284,5 +202,31 @@ public class EntityPacketRewriter1_16 {
                 wrapper.cancel();
             }
         });
+    }
+
+    @Override
+    protected void registerRewrites() {
+        filter().mapMetaType(Types1_16.META_TYPES::byId);
+        registerMetaTypeHandler(Types1_16.META_TYPES.itemType, Types1_16.META_TYPES.optionalBlockStateType, Types1_16.META_TYPES.particleType);
+        filter().type(EntityTypes1_16.ABSTRACT_MINECART).index(10).handler((metadatas, meta) -> {
+            int data = meta.value();
+            meta.setValue(protocol.getMappingData().getNewBlockStateId(data));
+        });
+        filter().type(EntityTypes1_16.ABSTRACT_ARROW).removeIndex(8);
+        filter().type(EntityTypes1_16.WOLF).index(16).handler((event, meta) -> {
+            byte mask = meta.value();
+            int angerTime = (mask & 0x02) != 0 ? Integer.MAX_VALUE : 0;
+            event.createExtraMeta(new Metadata(20, Types1_16.META_TYPES.varIntType, angerTime));
+        });
+    }
+
+    @Override
+    public void onMappingDataLoaded() {
+        mapTypes();
+    }
+
+    @Override
+    public EntityType typeFromId(int type) {
+        return EntityTypes1_16.getTypeFromId(type);
     }
 }
