+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_17_1to1_18/data/BlockEntityIds1_18.java
@@ -20,7 +20,7 @@ package com.viaversion.viaversion.protocols.v1_17_1to1_18.data;
 import com.viaversion.viaversion.api.Via;
 import java.util.Arrays;
 
-public final class BlockEntityIds {
+public final class BlockEntityIds1_18 {
 
     private static final int[] IDS = new int[14];
