+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_19_1to1_19_3/storage/NonceStorage1_19_3.java
@@ -20,5 +20,5 @@ package com.viaversion.viaversion.protocols.v1_19_1to1_19_3.storage;
 import com.viaversion.viaversion.api.connection.StorableObject;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-public record NonceStorage(byte @Nullable[] nonce) implements StorableObject {
+public record NonceStorage1_19_3(byte @Nullable[] nonce) implements StorableObject {
 }
