+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_19to1_19_1/data/ChatRegistry1_19_1.java
@@ -15,9 +15,20 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viaversion.protocols.v1_18_2to1_19.storage;
+package com.viaversion.viaversion.protocols.v1_19to1_19_1.data;
 
-import com.viaversion.viaversion.api.connection.StorableObject;
+import com.viaversion.nbt.tag.CompoundTag;
+import com.viaversion.viaversion.api.data.MappingDataLoader;
 
-public record NonceStorage(byte[] nonce) implements StorableObject {
+public class ChatRegistry1_19_1 {
+
+    private static final CompoundTag chatRegistry;
+
+    static {
+        chatRegistry = MappingDataLoader.INSTANCE.loadNBTFromFile("chat-registry-1.19.1.nbt");
+    }
+
+    public static CompoundTag chatRegistry() {
+        return chatRegistry.copy();
+    }
 }
