+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_19to1_19_1/storage/NonceStorage1_19_1.java
@@ -20,6 +20,6 @@ package com.viaversion.viaversion.protocols.v1_19to1_19_1.storage;
 import com.viaversion.viaversion.api.connection.StorableObject;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-public record NonceStorage(byte @Nullable [] nonce) implements StorableObject {
+public record NonceStorage1_19_1(byte @Nullable [] nonce) implements StorableObject {
 
 }
